Use proper type for return value of read(2)
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38165 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/gtk3/joy-unix.c
@@ -390,7 +390,7 @@ void old_joystick_init(void)
             /* calibration loop */
             for (j = 0; j < JOYCALLOOPS; j++) {
                 struct JS_DATA_TYPE js;
-                int status = read(ajoyfd[i], &js, JS_RETURN);
+                ssize_t status = read(ajoyfd[i], &js, JS_RETURN);
 
                 if (status != JS_RETURN) {
                     log_warning(joystick_log, "Error reading joystick device `%s'.", dev);
@@ -474,7 +474,7 @@ void old_joystick(void)
         } else
 #    endif
         if (joyport == JOYDEV_ANALOG_0 || joyport == JOYDEV_ANALOG_1) {
-            int status;
+            ssize_t status;
             struct JS_DATA_TYPE js;
             int ajoyport = joyport - JOYDEV_ANALOG_0;
